Add links based on content item
This adds a bunch of useful links to the popup.

- An Edit-link that leads to the publishing application. For some
applications we can infer a deeplink from the content_id (like
collections-publisher). For others we either link to a prefilled search
path (publisher) or link to the publishing app main page.
- Link to content-tagger (where anything can be tagged) and
search-admin (where you can see what raw information rummager holds)
- Links to the rendering app was already available in the popup, but we
now use the content item for this. This means that the link is also
available on pages that 404, 410 and 5XX. Previously this link was
based on a meta-tag which wouldn't be present on clobbered error pages.
- A similar link to the GitHub of the publishing application. In some
cases this needs some mapping, because the `publishing_app` in content
store isn't always the repo name. Make a PR if you find more exceptions.
- Link to the content schema on GitHub.
- A link to the brand new publishing-api debug page, which is initially
only available in development
(https://github.com/alphagov/publishing-api/pull/345).

The way we generate these links is by fetching the content-item from
the content-store when the user opens the popup. This means a slight
delay between opening and the extension populating the menu.