Add the ability to view smart answers GovSpeak version.
Here I have added a link to the GOV.UK chrome extension that simply adds a '.txt' to the end of the users current url.

This will then display the GovSpeak (Markdown) version of the page. This may not work for all urls and at the moment only
supported by SmartAnswers.

TODO:

* Need to get this tests working by faking the success json response.
* On the extension UI highlight the current link