Do not show GovSpeak on smartanswer landing pages
There is a known bug where the govspeak version of smartanswer landing pages
can only be viewed from the development environment.

There is a card in the backlog to fix this: https://trello.com/c/qd5C7qDn

Until the bug a fixed, do not display the "Display GovSpeak" link on smartanswer
landing pages.