Remove smart answer path parsing
The functionality removed here returns the "root" base paths for smart
answers. For example

/maternity-paternity-calculator/y/maternity

turns into

/maternity-paternity-calculator

This was added to generate the correct content store URLs, but breaks
other behaviour in the app, like being able to switch from feedback
to GOV.UK representation. It will be implemented differently in a
follow up commit.