Allow users to select an A/B bucket on Integration and Staging
Make the A/B test buckets interactive, so that a user can select the
version that they want to see. This works by appending a custom HTTP
header to the request, which will only work against origin servers - it
will be overridden by Varnish configuration if the request passes
through the CDN. A future commit will add the same functionality using a
cookie, which will work with the CDN.

Note that the selected A/B buckets are not persisted when the browser is
closed.