Allow users to select an A/B bucket on Production
When the user selects a different A/B testing bucket, update the value
of the relevant A/B cookie. This controls the user's A/B bucket on
environments with a CDN, such as Production.

The custom header is still appended to the request, so that non-CDN
environments still place the user in the correct bucket.