Display correct A/B buckets when popup is reloaded
This fixes a bug in which the A/B bucket values were not being displayed
correctly when the extension's popup was closed and re-opened. The
values were always being loaded from the metadata on the page, even when
the user had chosen different buckets.

The user's chosen buckets were being used correctly to set the headers
and cookie - the problem was just in the display.

Note that this fix still doesn't persist A/B buckets between browser
sessions in Dev/Integration/Staging. It just persists the buckets when
the popup is closed and opened.