Prevent popup from being re-rendered, since it resets click handlers
Previously, if there were content links, then the popup would be
rendered twice. Most of the click handlers were being set up again, but
not the new A/B bucket toggles.

This fix ensures that the popup is only rendered once when it is opened,
so the click handlers are never reset.