Preserve cookie path when user chooses an A/B bucket
This fixes a bug which caused the extension to create an extra cookie
named `ABTest-Example`, rather than updating the value of the original
cookie.

This is because the extension assumed that the cookie path is always the
root path, which isn't always true. This is fixed by copying the cookie
`path` attribute from the original cookie, which causes the browser to
overwrite the original rather than duplicate it.