Pick up new A/B tests without having to reload extension
Before, we were only initializing A/B tests the first time the popup
detected any A/B meta tags. This means that if a new A/B was released,
it would not appear in the popup until the user restarted Chrome or the
extension.

This would also be a problem if you switched between environments with
different A/B tests defined.