Used allowed variants when toggling A/B tests
We are allowing for variants to have different names than just A and B.
That means we need to update the extension to read the allowed variants
for a particular A/B test from the metatag, which will include them for
newer versions of `govuk_ab_testing`. In order to stay compatible with
older versions, we default to "A" and "B" as the bucket names in case
we can't find them in the metatag.