Add bookmarklet to highlight templates
When a content designer needs to edit smart answer content, they find
it difficult to find the relevant templates and partials that a
particular outcome or question uses.

This commits makes it possible for content designers to highlight the
templates and partials that a smart answer outcome or question has.

This has been done by providing a bookmarklet [1] that invokes the
linkToTemplatesOnGithub function.

[1] https://bl.ocks.org/floehopper/91b990d66bfae1dbdbaa88aa4772f1a9