Add function to outline GOV.UK components on page
We hope to encourage re-use of design elements that are documented in
the component guides [1][2] by highlighting where they exist in the
page.

The user can toggle this functionlity on from the extension popup,
which will outline all the components in the page with a label of what
it is called. Clicking within the outline will navigate the user to
its documentation.

[1] https://govuk-static.herokuapp.com/component-guide
[2] https://government-frontend.herokuapp.com/component-guide