Register click events for switching environments
In a previous release, the selector to attach click events was updated
to split apart internal and external links. However, the links that
changed the enviornments were missed and have since stopped working.

This has been resolved by adding an 'in-popup' class to all links in
the page that change the content in the current page.