Add functionality to toggle meta tags
Meta tags are shown in a big banner at the top of the page.

For now I've just copied the markup from an existing booklet.
The appearence varies a bit depending on which page you're on, but I
think this can be fixed later by someone who can write css.