Attempt to force a scrollbar when the extension doesn't fit on screen
Someone in the content team wants to use the extension but uses screen
magnification. This causes the extension window to fall off the screen.
Because the height is set you can't scroll.

This PR attempts to remedy the issue by setting the height of the
extension if the users' browser window is lower than 600 (the max
height of the popup).