Remove highlighting for Design System components
The gem can't differentiate global classes like `govuk-width-container`
from components in the design system, so it highlights a lot of things
incorrectly and links to non-existent pages on the design system site.

Removing this is not really a problem because we always wrap GOV.UK
Frontend components in our own component.