Adds contextual info when transforming links based on environment
I was previewing a piece of content on the draft stack when I used the
extension to view the content item. I was confused that this took me
to the live content item on GOV.UK, rather than the content item on
integration that I wanted.

Then I realised (when I navigated to the integration content item) that
it didn't contain the content I expected - and that's because the content
was still in draft, not published. I now understood why the extension
was unable to link me to the content store environment I wanted - it
doesn't exist.

Have now added some contextual information only when we override the
origin we redirect people to.