Delete unused Rakefile
The build script has been replaced by
https://github.com/alphagov/govuk-browser-extension/blob/9a5ae779e126b09fd303147d91c616c0fa527462/build.sh#L1

And the jasmine test script has been replaced by NPM:
https://github.com/alphagov/govuk-browser-extension/blob/7913847c4aa8b366826ada973264798cdb113ed2/.github/workflows/ci.yml#L36-L37

The presence of the Rakefile is causing issues for our CodeQL
integration in #196, which is detecting Ruby but then failing to
run the scan. It is hoped that removing the Rakefile will resolve
this.