Update Firefox add-on ID
This is the AOID provided by Firefox for our existing extension,
so let's remove the (randomly generated, unused) GUID that was
introduced in #214.
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/browser_specific_settings#firefox_gecko_properties

NB, it looks like it's fine for this ID to be public, as anyone
can inspect the IDs of their installed add-ons by visiting
`about:config` and searching for `extensions.webextensions.uuids`.

Trello: https://trello.com/c/F7Kv3G8j/3055-fix-govuk-browser-extension-release-process-for-multiple-browsers