Lookup in Mainstream using content id
We are going back to the changes in the PR we created originally
https://github.com/alphagov/govuk-browser-extension/pull/212

This change is coupled with the changes we have made in Mainstream
https://github.com/alphagov/publisher/pull/2456

These changes combined will support the old bootstrap ui and new design
system ui. Look up in mainstream will land user on publications page and
user will see same result to AS-IS for bootstrap ui.